Unify metrics and forwarding tests
